Fix `Handler` and `HandlerM` and `EndoM`
There were two things wrong here that were not caught for a long time.

First, the type of `Handler` was wrong because it was folding things
"backwards".  For example:

    >>> fold (handles Control.Lens.folded list) ["ABC", "DEF"]
    "FEDCBA"

This also caused a space leak when folding larger collections.

The problem is easy to fix, just add a `Dual` to the type of `Handler` to switch
the order in which step functions are composed.  This fixes the order of the
list and now the fold runs in constant space:

    >>> fold (handles Control.Lens.folded list) ["ABC", "DEF"]
    "ABCDEF"

The second problem was that the `EndoM` `Monoid` was backwards (meaning that it
did not compose kleisli arrows in the same order that `Endo` composed
functions).  I'm guessing that this was because I noticed the same backwards
folding issue when adding `EndoM` and reversed the order to fix the problem.

What this means was that for a long time `handles` was folding things in the
wrong order and `handlesM` was folding things in right order (due to `EndoM`
already composing things backwards).  I decided to fix both, though, in order
for things to be more consistent.  In other words, I:

* flipped the fold order of `handles` by adding `Dual` to `Handler`
* changed the composition order in `EndoM` to match `Endo`
* then I canceled out the previous change by adding `Dual` to `HandlerM`, too

Now both `handles` and `handlesM` fold things in the correct order, and `EndoM`
is consistent with `Endo` and the `Handles` and `HandlerM` types both use
`Dual` now.