Add `folded`.  Fixes #61
This is a utility optic that is designed to be used with `handles` and
`handlesM`.  Normally users could get this through the `lens-family-core` or
the `lens` library, but I've gotten the request for this functionality so many
times that I figure I should just export it directly from this library for users
who don't want to depend on a lens library.